+++ main.go
@@ -24,7 +24,7 @@ import (
 
 // Build Infos
 var (
-	Version = "3.6.0"
+	Version = "3.6.1"
 )
 
 // Flags